+++ MGen/Generators/GenCF1.cpp
@@ -711,7 +711,7 @@ int CGenCF1::FailMelodyHarm(vector<int> &pc) {
 		}
 		// Shuffle
 		if (task == tEval) {
-			unsigned seed = CGLib::time();
+			long long seed = CGLib::time();
 			::shuffle(hm[ls].begin(), hm[ls].end(), default_random_engine(seed));
 		}
 	}
@@ -2194,7 +2194,7 @@ void CGenCF1::ScanLeft(vector<int> &cc, int &finished) {
 void CGenCF1::BackWindow(vector<int> &cc) {
 	// Show best rejected variant
 	if (best_rejected) {
-		int time = CGLib::time();
+		long long time = CGLib::time();
 		int rc = (time - accept_time) / best_rejected;
 		if (debug_level > 2) {
 			CString st;
@@ -3022,7 +3022,7 @@ void CGenCF1::RandomSWA()
 // Do not calculate dpenalty (dp = 0). Calculate dpenalty (dp = 1).
 void CGenCF1::SWA(int i, int dp) {
 	CString st;
-	int time_start = CGLib::time();
+	long long time_start = CGLib::time();
 	s_len = 1;
 	// Save source rpenalty
 	float rpenalty_source = rpenalty_cur;
@@ -3114,7 +3114,7 @@ void CGenCF1::SWA(int i, int dp) {
 		}
 	}
 	// Log
-	int time_stop = CGLib::time();
+	long long time_stop = CGLib::time();
 	CString est;
 	CString stuck_st = GetStuck();
 	est.Format("Finished SWA%d #%d: rp %.0f from %.0f, dp %.0f, cnum %ld (in %d ms): %s", 
@@ -3352,7 +3352,7 @@ void CGenCF1::Generate()
 		vector<Color> color2(t_generated);
 		for (int i = 0; i < accepted; ++i) ci[i] = i;
 		// Shuffled indexes
-		unsigned seed = CGLib::time();
+		long long seed = CGLib::time();
 		::shuffle(ci.begin(), ci.end(), default_random_engine(seed));
 		// Swap
 		int s1, s2;
